+++ src/base/b-virtual-scroll/test/runners/slots/render-next.js
@@ -45,6 +45,7 @@ module.exports = (page) => {
 
 	beforeEach(async () => {
 		await h.utils.reloadAndWaitForIdle(page);
+		await h.component.waitForComponent(page, '#root-component');
 
 		await page.evaluate(() => {
 			const dummy = document.querySelector('#dummy-component');